vendorsystem/requirement.txt
............................Outside the virtualenv................................
appdirs==1.4.3
astroid==2.3.3
autopep8==1.5.2
colorama==0.4.3
distlib==0.3.0
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.4.4
six==1.14.0
virtualenv==20.0.18
wrapt==1.11.2
............................Inside the virtualenv......................................
asgiref==3.2.7
Babel==2.8.0
Django==3.0.5
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
phonenumbers==8.12.3
pytz==2019.3
sqlparse==0.3.1